Set up timer for delay-from-last action object strategy

+++ domain/src/main/kotlin/com/ringoid/domain/action_storage/TriggerStrategy.kt
@@ -4,8 +4,8 @@ sealed class TriggerStrategy
 
 object Immediate : TriggerStrategy()
 class CountFromLast(val count: Int = LIKE_COUNT_ON_TRIGGER) : TriggerStrategy()
-class DelayFromLast(val delay: Int = LIKE_DELAY_ON_TRIGGER) : TriggerStrategy()  // in seconds
+class DelayFromLast(val delay: Long = LIKE_DELAY_ON_TRIGGER) : TriggerStrategy()  // in seconds
 
 const val LIKE_COUNT_ON_TRIGGER = 5
-const val LIKE_DELAY_ON_TRIGGER = 5  // in seconds
-const val VIEW_DELAY_ON_TRIGGER = 10
+const val LIKE_DELAY_ON_TRIGGER = 5L  // in seconds
+const val VIEW_DELAY_ON_TRIGGER = 10L
